🔥 chore: move project service handlers to handlers directory
