+++ src/services/ProjectService/handlers/removeProject.ts
@@ -1,11 +1,11 @@
 import { startSession } from "mongoose";
 
-import Config from "../../config";
-import { User } from "../../models/User";
-import { Repo } from "../../models/Repo";
-import ProjectService from ".";
-import { createDeploymentDebug } from "../../utils/createDebug";
-import { DeploymentError } from "../../utils/errors";
+import Config from "../../../config";
+import { User } from "../../../models/User";
+import { Repo } from "../../../models/Repo";
+import ProjectService from "..";
+import { createDeploymentDebug } from "../../../utils/createDebug";
+import { DeploymentError } from "../../../config/errors";
 
 const removeProject = async (service: ProjectService, next: Function) => {
   const debug = createDeploymentDebug(Config.CLIENT_OPTIONS.debug);
