+++ src/services/ProjectService/handlers/setGithubInfo.ts
@@ -1,9 +1,9 @@
-import GithubClient from "../GithubClient";
-import { DeploymentError } from "../../utils/errors";
-import { createDeploymentDebug } from "../../utils/createDebug";
-import Config from "../../config";
+import GithubClient from "../../GithubClient";
+import { DeploymentError } from "../../../config/errors";
+import { createDeploymentDebug } from "../../../utils/createDebug";
+import Config from "../../../config";
 
-import ProjectService from "./";
+import ProjectService from "..";
 
 const setGithubInfo = async (service: ProjectService, next: Function) => {
   const debug = createDeploymentDebug(Config.CLIENT_OPTIONS.debug);
